+++ src/programming-learning-tool/src/components/Collapsible.vue
