+++ src/programming-learning-tool/src/components/GroupCollapsible.vue
